correction des commentaire de la pr
