+++ bff/BffGreenWaiter/src/main/java/fr/green/BffGreenWaiter/allergen/controller/AllergenController.java
@@ -1,7 +1,7 @@
 package fr.green.BffGreenWaiter.allergen.controller;
 
 import fr.green.BffGreenWaiter.allergen.service.AllergenService;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,13 +10,11 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/allergen")
 public class AllergenController {
-    @Autowired
-    private AllergenService allergenService;
-
-    public AllergenController() {}
-
+    private final AllergenService allergenService;
+    
     @GetMapping("/getAllergen/{id}")
     public List<String> getAllergen(@PathVariable String  id) {
         return allergenService.getAllergensById(id);
