+++ bff/BffGreenWaiter/src/main/java/fr/green/BffGreenWaiter/allergen/model/Allergen.java
@@ -1,29 +1,17 @@
 package fr.green.BffGreenWaiter.allergen.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.Setter;
+
 import java.util.List;
 
-
+@Getter
+@Setter
 public class Allergen {
 
     @JsonProperty("_id")
     private String id;
     private List<String> allergens;
 
-    // Getters et Setters
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public List<String> getAllergens() {
-        return allergens;
-    }
-
-    public void setAllergens(List<String> allergens) {
-        this.allergens = allergens;
-    }
 }
