+++ frontend/src/components/menu/menu-item/menu-item.tsx
@@ -3,8 +3,7 @@ import React, { useEffect, useState } from 'react';
 import './menu-item.scss';
 import IconButton from '../../common/icon-button/icon-button.tsx';
 import AllergenPopup from '../allergen-popup/allergen-popup.tsx';
-import config from '../../../config';
-import data from '../../../assets/allergen.json';
+import getAllergens from '../../../services/service-allergen.ts';
 
 type MenuItemProps = {
   item: Item;
@@ -22,33 +21,6 @@ export default function MenuItem({ item, onClick, className }: Readonly<MenuItem
     setIsOpen(true);
   };
   const close = () => setIsOpen(false);
-
-  // ça pourrait être chargé au moment de la creation des items, pour l'instant ça n'a pas été fait mais au moment de la recup des items il faudra le faire
-  const getAllergens = async (id: string): Promise<string[]> =>
-    config.bffFlag ? await getAllergensBff(id) : getAllergensNoBff(id);
-
-  const getAllergensBff = async (id: string): Promise<string[]> => {
-    try {
-      const res = await fetch(`${config.bffApi}allergen/getAllergen/${id}`, {
-        method: 'GET',
-        headers: { Accept: 'application/json' },
-      });
-      if (!res.ok) {
-        console.error('Erreur HTTP allergènes', res.status);
-        return [];
-      }
-      const payload = await res.json();
-      return Array.isArray(payload) ? payload : (payload?.allergens ?? []);
-    } catch (e) {
-      console.error('Erreur réseau allergènes', e);
-      return [];
-    }
-  };
-
-  const getAllergensNoBff = (id: string): string[] => {
-    const itemAllergen = data.find((a) => a._id === id);
-    return itemAllergen ? itemAllergen.allergens : [];
-  };
 
   useEffect(() => {
     let mounted = true;
